+++ GDX/Runtime/IO/Compression/TarFile.cs
@@ -18,11 +18,15 @@ namespace GDX.IO.Compression
         ///     The path to the directory in which to place the extracted files, specified as a
         ///     relative or absolute path. A relative path is interpreted as relative to the current working directory.
         /// </param>
-        /// <param name="forceGZipDataFormat">Enforce inflating the file via a <see cref="GZipStream"/>.</param>
-        public static void ExtractToDirectory(string sourceArchiveFileName, string destinationDirectoryName, bool forceGZipDataFormat = false)
+        /// <param name="forceGZipDataFormat">Enforce inflating the file via a <see cref="GZipStream" />.</param>
+        public static void ExtractToDirectory(string sourceArchiveFileName, string destinationDirectoryName,
+            bool forceGZipDataFormat = false)
         {
+            const int readBufferSize = 4096;
+
             // We need to handle the gzip first before we address the archive itself
-            if (forceGZipDataFormat || sourceArchiveFileName.EndsWith(".gz", StringComparison.InvariantCultureIgnoreCase))
+            if (forceGZipDataFormat ||
+                sourceArchiveFileName.EndsWith(".gz", StringComparison.InvariantCultureIgnoreCase))
             {
                 using FileStream stream = File.OpenRead(sourceArchiveFileName);
                 using GZipStream gzip = new GZipStream(stream, CompressionMode.Decompress);
@@ -30,19 +34,18 @@ namespace GDX.IO.Compression
 
                 // Loop through the stream
                 int readByteCount;
-                byte[] readBuffer = new byte[4096];
+                byte[] readBuffer = new byte[readBufferSize];
                 do
                 {
-                    readByteCount = gzip.Read(readBuffer, 0, 4096);
+                    readByteCount = gzip.Read(readBuffer, 0, readBufferSize);
                     memoryStream.Write(readBuffer, 0, readByteCount);
-                } while (readByteCount == 4096);
+                } while (readByteCount == readBufferSize);
 
                 memoryStream.Seek(0, SeekOrigin.Begin);
                 ExtractStream(memoryStream, destinationDirectoryName);
             }
             else
             {
-
                 using FileStream fileStream = File.OpenRead(sourceArchiveFileName);
                 ExtractStream(fileStream, destinationDirectoryName);
             }
@@ -50,66 +53,60 @@ namespace GDX.IO.Compression
 
 
         /// <summary>
-        ///     Extract a tarball to the <paramref name="destinationDirectoryName"/>.
+        ///     Extract a tarball to the <paramref name="destinationDirectoryName" />.
         /// </summary>
-        /// <param name="sourceStream">The <see cref="Stream"/> which to extract from.</param>
+        /// <param name="sourceStream">The <see cref="Stream" /> which to extract from.</param>
         /// <param name="destinationDirectoryName">Output directory to write the files.</param>
         private static void ExtractStream(Stream sourceStream, string destinationDirectoryName)
         {
-            byte[] buffer = new byte[100];
+            const int readBufferSize = 100;
+            const int contentOffset = 512;
+            byte[] readBuffer = new byte[readBufferSize];
             while (true)
             {
-                sourceStream.Read(buffer, 0, 100);
-                string currentName = Encoding.ASCII.GetString(buffer).Trim('\0');
+                sourceStream.Read(readBuffer, 0, readBufferSize);
+                string currentName = Encoding.ASCII.GetString(readBuffer).Trim('\0');
 
                 if (string.IsNullOrWhiteSpace(currentName))
                 {
                     break;
                 }
 
+                string destinationFilePath = Path.Combine(destinationDirectoryName, currentName);
                 sourceStream.Seek(24, SeekOrigin.Current);
-                sourceStream.Read(buffer, 0, 12);
-
-                long fileSize = Convert.ToInt64(Encoding.UTF8.GetString(buffer, 0, 12).Trim('\0').Trim(), 8);
-
+                sourceStream.Read(readBuffer, 0, 12);
+                long fileSize = Convert.ToInt64(Encoding.UTF8.GetString(readBuffer, 0, 12).Trim('\0').Trim(), 8);
                 sourceStream.Seek(376L, SeekOrigin.Current);
 
-                string output = Path.Combine(destinationDirectoryName, currentName);
-                if (!Directory.Exists(Path.GetDirectoryName(output)))
+                // Do we need to make a directory?
+                string parentDirectory = Path.GetDirectoryName(destinationFilePath);
+                if (!Directory.Exists(parentDirectory))
                 {
-                    string directoryPath = Path.GetDirectoryName(output);
-                    Directory.CreateDirectory(directoryPath);
-
-                    //TODO: Need to handle permissions on nearly created
-                    // DirectoryInfo directory = new DirectoryInfo(directoryPath);
-                    // DirectorySecurity security = directory.GetAccessControl();
-                    //
-                    // security.AddAccessRule(new FileSystemAccessRule(@"MYDOMAIN\JohnDoe",
-                    //     FileSystemRights.Modify,
-                    //     AccessControlType.Deny));
-                    //
-                    // directory.SetAccessControl(security);
+                    // ReSharper disable once AssignNullToNotNullAttribute
+                    DirectoryInfo directoryInfo = Directory.CreateDirectory(parentDirectory);
+                    directoryInfo.Attributes &= ~FileAttributes.ReadOnly;
                 }
 
-                if (!currentName.Equals("./", StringComparison.InvariantCulture))
+                // Don't try to make directories as files
+                if (!currentName.Equals("./", StringComparison.InvariantCulture) &&
+                    !currentName.EndsWith("/") &&
+                    !currentName.EndsWith("\\"))
                 {
-                    using FileStream newFileStream = File.Open(output, FileMode.OpenOrCreate, FileAccess.Write);
+                    using FileStream newFileStream =
+                        File.Open(destinationFilePath, FileMode.OpenOrCreate, FileAccess.Write);
                     byte[] fileContentBuffer = new byte[fileSize];
                     int newFileContentBufferLength = fileContentBuffer.Length;
-
                     sourceStream.Read(fileContentBuffer, 0, newFileContentBufferLength);
                     newFileStream.Write(fileContentBuffer, 0, newFileContentBufferLength);
                 }
 
-                long pos = sourceStream.Position;
-
-                long offset = 512 - pos % 512;
-                if (offset == 512)
+                long nextOffset = contentOffset - sourceStream.Position % contentOffset;
+                if (nextOffset == contentOffset)
                 {
-                    offset = 0;
+                    nextOffset = 0;
                 }
 
-                sourceStream.Seek(offset, SeekOrigin.Current);
+                sourceStream.Seek(nextOffset, SeekOrigin.Current);
             }
         }
     }